backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
openai-whisper
ffmpeg-python
static-ffmpeg
rapidfuzz
requests
beautifulsoup4
lyricsgenius
aiofiles
python-dotenv
httpx
